refactor(quota): extract OpenCode Go and Ollama Cloud scraping helpers
