+++ src/shared/constants/providers.ts
@@ -1,5 +1,3 @@
-// Provider definitions
-
 /**
  * Service kind — declarative tag for what a provider can do beyond basic LLM chat.
  * Affects UI filtering and playground routing; does not influence request routing.
